Add imitation model engine

+++ src/main/java/imitationmodel/Distribution.java
@@ -11,5 +11,5 @@ public abstract class Distribution {
         this.distributionType = distributionType;
     }
 
-    public abstract int calculateNext();
+    public abstract Double calculateNext();
 }
